requirements.txt
pytube==15.0.0
youtube-transcript-api==0.6.1
autogen>=0.2.16
sentence-transformers==2.2.2
torch>=1.9.0
transformers>=4.21.0
scikit-learn>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.64.0
yt-dlp>=2023.12.30
tabulate>=0.9.0
pandas>=1.5.0 